requirements.txt
Django==2.0.5
djangorestframework==3.8.0
psycopg2-binary==2.8.3
pandas
djangorestframework-camel-case==0.2.0
pillow
django-cors-headers
